Suppress clang compiler warnings.

+++ examples/viewer/viewer.c
@@ -63,7 +63,7 @@ static void CalcNormal(float N[3], float v0[3], float v1[3], float v2[3]) {
 
   len2 = N[0] * N[0] + N[1] * N[1] + N[2] * N[2];
   if (len2 > 0.0f) {
-    float len = sqrt(len2);
+    float len = (float)sqrt((double)len2);
 
     N[0] /= len;
     N[1] /= len;
